requirements.txt
pygame
graphviz
fpdf2
Pillow